Decouple preview transform table from define transform
